+++ TP3/src/Exemplaire.h
@@ -10,23 +10,23 @@
 class Exemplaire
 {
 public:
-  Exemplaire(int nbPoints);
-  Exemplaire(int nbPoints, std::vector<int>& types, std::vector<int>& maxSentiers, std::vector<std::vector<double> >& couts);
-  ~Exemplaire();
-  void setNbPoints(int nbPoints);
-  void setTypes(std::vector<int>& types);
-  void setMaxSentiers(std::vector<int>& maxSentiers);
-  void setCouts(std::vector<std::vector<double> >& couts);
-  int getNbPoints();
-  std::vector<int> getTypes();
-  std::vector<int> getMaxSentiers();
-  std::vector<std::vector<double> > getCouts();
+    Exemplaire(int nbPoints);
+    Exemplaire(int nbPoints, std::vector<int>& types, std::vector<int>& maxSentiers, std::vector<std::vector<int> >& couts);
+    ~Exemplaire();
+    void setNbPoints(int nbPoints);
+    void setTypes(std::vector<int>& types);
+    void setMaxSentiers(std::vector<int>& maxSentiers);
+    void setCouts(std::vector<std::vector<int> >& couts);
+    int getNbPoints();
+    std::vector<int> getTypes();
+    std::vector<int> getMaxSentiers();
+    std::vector<std::vector<int> > getCouts();
 
 private:
-  int nbPoints; // nombre de points dans le parc
-  std::vector<int> types; // vecteur indiquant le type de chaque point
-  std::vector<int> maxSentiers; // vecteur indiquant pour chaque point son maximum de sentiers incidents autorise
-  std::vector<std::vector<double> > couts; // matrice des couts entre chaque point
+    int nbPoints; // nombre de points dans le parc
+    std::vector<int> types; // vecteur indiquant le type de chaque point
+    std::vector<int> maxSentiers; // vecteur indiquant pour chaque point son maximum de sentiers incidents autorise
+    std::vector<std::vector<int> > couts; // matrice des couts entre chaque point
 };
 
 #endif
