+++ TP3/src/solveur.h
@@ -8,11 +8,11 @@
 
 Solution* trouverPremiereSolution(Exemplaire& e);
 
-int minKey(std::vector<double>& key, std::vector<int>& mstSet, int nbPoints);
+int minKey(std::vector<int>& key, std::vector<int>& mstSet, int nbPoints);
 
-void printMST(std::vector<int>& parent, int n, std::vector<std::vector<double> >& graph, int nbPoints);
+void printMST(std::vector<int>& parent, int n, std::vector<std::vector<int> >& graph, int nbPoints);
 
-std::vector<int> primMST(std::vector<std::vector<double> >& graph, int nbPoints);
+std::vector<int> primMST(std::vector<std::vector<int> >& graph, int nbPoints);
 
 void amelioration(Solution &s, Exemplaire &e, bool afficherSol, bool afficherTmp, clock_t debut);
